requirements.txt
asgiref==3.10.0
contourpy==1.3.3
cycler==0.12.1
Django==5.2.8
fonttools==4.60.1
joblib==1.5.2
kiwisolver==1.4.9
liac-arff==2.5.0
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
gunicorn==23.0.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9